backend/requirements.txt
fastapi==0.110.3
uvicorn==0.30.6
sqlalchemy==2.0.49
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.9
email-validator==2.1.2
python-dotenv==1.0.1
